+++ commands/msgview/open-link.go
@@ -2,6 +2,8 @@ package msgview
 
 import (
 	"errors"
+	"fmt"
+	"net/url"
 
 	"git.sr.ht/~rjarry/aerc/commands"
 	"git.sr.ht/~rjarry/aerc/lib"
@@ -30,12 +32,17 @@ func (OpenLink) Complete(aerc *widgets.Aerc, args []string) []string {
 }
 
 func (OpenLink) Execute(aerc *widgets.Aerc, args []string) error {
-	if len(args) != 2 {
-		return errors.New("Usage: open-link <url>")
+	if len(args) < 2 {
+		return errors.New("Usage: open-link <url> [program [args...]]")
 	}
+	u, err := url.Parse(args[1])
+	if err != nil {
+		return err
+	}
+	mime := fmt.Sprintf("x-scheme-handler/%s", u.Scheme)
 	go func() {
 		defer log.PanicHandler()
-		if err := lib.XDGOpen(args[1]); err != nil {
+		if err := lib.XDGOpenMime(args[1], mime, args[2:]); err != nil {
 			aerc.PushError("open-link: " + err.Error())
 		}
 	}()